FYP/requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==2.3.3
Flask-Mail==0.9.1
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
mailjet-rest==1.4.0
MarkupSafe==3.0.2
mysql-connector-python==8.1.0
packaging==25.0
pillow==11.3.0
protobuf==4.21.12
PyMySQL==1.1.0
reportlab==4.0.4
requests==2.32.4
urllib3==2.5.0
Werkzeug==2.3.7 